Adding Chapter 6 Training method into the utils class. Change-Id: I57b900404f592c85a4ef513843a064a7d2ed216c
